+++ hosts/phi/configuration.nix
@@ -22,7 +22,11 @@
   services.xserver.displayManager.gdm.autoSuspend = false;
 
   sound.enable = true;
-  hardware.pulseaudio.enable = true;
+  hardware.pulseaudio.enable = false;
+  services.pipewire.enable = true;
+  services.pipewire.alsa.enable = true;
+  services.pipewire.alsa.support32Bit = true;
+  services.pipewire.pulse.enable = true;
 
   # Enable FreeSync
   services.xserver.deviceSection = ''